+++ src/command_modules/azure-cli-lab/azure/cli/command_modules/lab/_help.py
@@ -130,6 +130,29 @@ helps['lab vm apply-artifacts'] = """
                 - name: --artifacts
                   short-summary: JSON encoded array of artifacts to be applied. Use @{file} to load from a file
             """
+helps['lab vm claim'] = """
+            type: command
+            short-summary: Claim a specific virtual machine or any available from the Lab.
+            parameters:
+                - name: --ids
+                  short-summary: Space separated list of VM IDs to claim.
+                - name: --resource-group -g
+                  short-summary: Name of lab's resource group
+                - name: --lab-name
+                  short-summary: Name of the Lab
+                - name: --name -n
+                  short-summary: Name of the virtual machine
+            examples:
+                - name: Claim any available virtual machine in the lab.
+                  text: >
+                    az lab vm claim -g MyRG --lab-name MyLab
+                - name: Claim a specific virtual machine in the lab.
+                  text: >
+                    az lab vm claim -g MyRG --lab-name MyLab --name MyVM
+                - name: Claim multiple virtual machines in the lab using --ids.
+                  text: >
+                    az lab vm claim --ids /subscriptions/{SubID}/resourcegroups/{MyRG}/providers/microsoft.devtestlab/labs/{MyLab}/virtualmachines/{MyVM1} /subscriptions/{SubID}/resourcegroups/{MyRG}/providers/microsoft.devtestlab/labs/{MyLab}/virtualmachines/{MyVM2}
+            """
 helps['lab custom-image'] = """
             type: group
             short-summary: Commands to manage custom images of a DevTest Lab.