+++ src/command_modules/azure-cli-monitor/azure/cli/command_modules/monitor/custom.py
@@ -6,6 +6,7 @@ from __future__ import print_function
 import datetime
 import os
 from azure.cli.core.util import get_file_json, CLIError
+from azure.mgmt.monitor.models import ConditionOperator, TimeAggregationOperator
 
 # 1 hour in milliseconds
 DEFAULT_QUERY_TIME_RANGE = 3600000
@@ -13,14 +14,135 @@ DEFAULT_QUERY_TIME_RANGE = 3600000
 # ISO format with explicit indication of timezone
 DATE_TIME_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
 
+# region Alerts
 
-def list_metric_definitions(client, resource_id, metric_names=None):
+operator_map = {
+    '>': ConditionOperator.greater_than.value,
+    '>=': ConditionOperator.greater_than_or_equal.value,
+    '<': ConditionOperator.less_than,
+    '<=': ConditionOperator.less_than_or_equal
+}
+
+
+aggregation_map = {
+    'avg': TimeAggregationOperator.average.value,
+    'min': TimeAggregationOperator.minimum.value,
+    'max': TimeAggregationOperator.maximum.value,
+    'total': TimeAggregationOperator.total.value,
+    'last': TimeAggregationOperator.last.value
+}
+
+
+def _parse_actions(actions):
+    """ Actions come in as a combined list. This method separates the webhook actions into a
+        separate collection and combines any number of email actions into a single email collection
+        and a single value for `email_service_owners`. If any email action contains a True value
+        for `send_to_service_owners` then it is assumed the entire value should be True. """
+    from azure.mgmt.monitor.models import RuleEmailAction, RuleWebhookAction
+    actions = actions or []
+    email_service_owners = None
+    webhooks = [x for x in actions if isinstance(x, RuleWebhookAction)]
+    custom_emails = set()
+    for action in actions:
+        if isinstance(action, RuleEmailAction):
+            if action.send_to_service_owners:
+                email_service_owners = True
+            custom_emails = custom_emails | set(action.custom_emails)
+    return list(custom_emails), webhooks, email_service_owners
+
+
+def _parse_action_removals(actions):
+    """ Separates the combined list of keys to remove into webhooks and emails. """
+    flattened = list(set([x for sublist in actions for x in sublist]))
+    emails = []
+    webhooks = []
+    for item in flattened:
+        if item.startswith('http://') or item.startswith('https://'):
+            webhooks.append(item)
+        else:
+            emails.append(item)
+    return emails, webhooks
+
+
+def create_metric_rule(client, resource_group_name, rule_name, target, condition,
+                       description=None, disabled=False, location=None, tags=None,
+                       email_service_owners=False, actions=None):
+    from azure.mgmt.monitor.models import AlertRuleResource, RuleEmailAction, RuleWebhookAction
+    condition.data_source.resource_uri = target
+    custom_emails, webhooks, _ = _parse_actions(actions)
+    actions = [RuleEmailAction(email_service_owners, custom_emails)] + (webhooks or [])
+    rule = AlertRuleResource(location, rule_name, not disabled,
+                             condition, tags, description, actions)
+    return client.create_or_update(resource_group_name, rule_name, rule)
+
+
+def update_metric_rule(instance, target=None, condition=None, description=None, enabled=None,
+                       metric=None, operator=None, threshold=None, aggregation=None, period=None,
+                       tags=None, email_service_owners=None, add_actions=None, remove_actions=None):
+    # Update general properties
+    if description is not None:
+        instance.description = description
+    if enabled is not None:
+        instance.is_enabled = enabled
+    if tags is not None:
+        instance.tags = tags
+
+    # Update conditions
+    if condition is not None:
+        target = target or instance.condition.data_source.resource_uri
+        instance.condition = condition
+
+    if metric is not None:
+        instance.condition.data_source.metric_name = metric
+    if operator is not None:
+        instance.condition.operator = operator_map[operator]
+    if threshold is not None:
+        instance.condition.threshold = threshold
+    if aggregation is not None:
+        instance.condition.time_aggregation = aggregation_map[aggregation]
+    if period is not None:
+        instance.condition.window_size = period
+
+    if target is not None:
+        instance.condition.data_source.resource_uri = target
+
+    # Update actions
+    emails, webhooks, curr_email_service_owners = _parse_actions(instance.actions)
+
+    # process removals
+    if remove_actions is not None:
+        removed_emails, removed_webhooks = _parse_action_removals(remove_actions)
+        emails = [x for x in emails if x not in removed_emails]
+        webhooks = [x for x in webhooks if x.service_uri not in removed_webhooks]
+
+    # process additions
+    if add_actions is not None:
+        added_emails, added_webhooks, _ = _parse_actions(add_actions)
+        emails = list(set(emails) | set(added_emails))
+        webhooks = webhooks + added_webhooks
+
+    # Replace the existing actions array. This potentially restructures rules that were created
+    # via other methods (Portal, ARM template). However, the functionality of these rules should
+    # be the same.
+    from azure.mgmt.monitor.models import RuleEmailAction
+    if email_service_owners is None:
+        email_service_owners = curr_email_service_owners
+    actions = [RuleEmailAction(email_service_owners, emails)] + webhooks
+    instance.actions = actions
+
+    return instance
+
+# endregion
+
+
+# pylint: disable=unused-argument
+def list_metric_definitions(client, resource, resource_group_name=None, metric_names=None):
     '''Commands to manage metric definitions.
     :param str resource_id: The identifier of the resource
     :param str metric_names: The list of metric names
     '''
     odata_filter = _metric_names_filter_builder(metric_names)
-    metric_definitions = client.list(resource_id, filter=odata_filter)
+    metric_definitions = client.list(resource, filter=odata_filter)
     return list(metric_definitions)
 
 
@@ -34,7 +156,8 @@ def _metric_names_filter_builder(metric_names=None):
     return ' or '.join(filters)
 
 
-def list_metrics(client, resource_id, time_grain,
+# pylint: disable=unused-argument
+def list_metrics(client, resource, time_grain, resource_group_name=None,
                  start_time=None, end_time=None, metric_names=None):
     '''Lists the metric values for a resource.
     :param str resource_id: The identifier of the resource
@@ -49,7 +172,7 @@ def list_metrics(client, resource_id, time_grain,
     :param str metric_names: The space separated list of metric names
     '''
     odata_filter = _metrics_odata_filter_builder(time_grain, start_time, end_time, metric_names)
-    metrics = client.list(resource_id, filter=odata_filter)
+    metrics = client.list(resource, filter=odata_filter)
     return list(metrics)