+++ src/azure-cli-core/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+unreleased
+^^^^^^^^^^^^^^^^^^
+* Command paths are no longer case sensitive.
+* Certain boolean-type parameters are no longer case sensitive.
+
 2.0.6 (2017-05-09)
 ^^^^^^^^^^^^^^^^^^
 * RP Auto-Reg: capture missing subscription registration error on LRO (#3268)